feat: 객체의 역할에 맞게 입력값 검증 로직 내부로 변경
